% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effect_commands}
\alias{effect_commands}
\title{Effect commands (data.frame)}
\format{
\code{effect_commands} is a \code{data.frame} with three columns:
\itemize{
\item \code{Code}: a \code{character} of pseudo code used to achieve a certain effect
in ProTracker
\item \code{Effect}: Name of the effect.
\item \code{Description}: Description of the effect.
}

It is used for documentation and reference purposes.
}
\description{
ProTracker uses codes to achieve certain effects or jump to a
specific position (see \code{vignette("effect_commands")}).
}
\examples{
data("effect_commands")
}
