% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProTrackR2-package.R
\docType{package}
\name{ProTrackR2-package}
\alias{ProTrackR2}
\alias{ProTrackR2-package}
\title{ProTrackR2: Manipulate and Play 'ProTracker' Modules}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'ProTracker' is a popular music tracker to sequence music on a Commodore Amiga machine. This package offers the opportunity to import, export, manipulate and play 'ProTracker' module files. Even though the file format could be considered archaic, it still remains popular to this date. This package intends to contribute to this popularity and therewith keeping the legacy of 'ProTracker' and the Commodore Amiga alive. This package is the successor of 'ProTrackR' providing better performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/ProTrackR2/}
  \item \url{https://github.com/pepijn-devries/ProTrackR2}
  \item Report bugs at \url{https://github.com/pepijn-devries/ProTrackR2/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Authors:
\itemize{
  \item Olav Sørensen (Developer of the ProTracker 2.3d clone)
}

}
\keyword{internal}
