% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{pt2_demo}
\alias{pt2_demo}
\title{Path to demonstration ProTracker module file}
\usage{
pt2_demo()
}
\value{
Returns a string with a path of a ProTracker module file
}
\description{
A path to demonstration ProTracker module file. It can be used for
demonstration purposes.
}
\author{
Pepijn de Vries
}
