% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{pt2_new_mod}
\alias{pt2_new_mod}
\title{Create a new (empty) ProTracker module}
\usage{
pt2_new_mod(name, ...)
}
\arguments{
\item{name}{Name for the new module. It will be truncated if longer than 20 characters.}

\item{...}{Ignored}
}
\value{
A \code{pt2mod} class module, with no samples and one empty pattern.
}
\description{
Creates an empty ProTracker module, it is returned as a \code{pt2mod} class object.
}
\examples{
mod <- pt2_new_mod("my_song")
}
\author{
Pepijn de Vries
}
