% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pt2_new_pattern}
\alias{pt2_new_pattern}
\title{Create a new ProTracker pattern}
\usage{
pt2_new_pattern(..., compact = TRUE)
}
\arguments{
\item{...}{Currently ignored}

\item{compact}{Should the pattern be formatted using a compact notation (as used for
file storage), or a none-compact format as used by the player? This can be
set with the \code{compact} argument.}
}
\value{
Returns a new clean \code{pt2pat} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Creates a new ProTracker pattern, consisting
of four channels and 64 rows.
}
\author{
Pepijn de Vries
}
