% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{pt2_note}
\alias{pt2_note}
\title{Extract a note from a ProTracker module}
\usage{
pt2_note(x, ...)
}
\arguments{
\item{x}{An object of class \code{pt2_cell}, which can be extracted
from a pattern table with \code{\link[=pt2_cell]{pt2_cell()}}.}

\item{...}{Ignored}
}
\value{
Returns a string representing the note's key.
}
\description{
Gets note information from a cell in a pattern table in a ProTracker Module.
}
\details{
A string representing the note's key is returned by the function.
The first letter indicates the position of the note in the
\href{https://en.wikipedia.org/wiki/Diatonic_scale}{diatonic scale}.
The second character indicates if it is a sharp key (with a hash
symbol, and a dash if it is not). The third character indicates
the octave of the note. In ProTracker allowed notes range from
\code{"C-1"} to \code{"B-3"}.
}
\examples{
mod <- pt2_read_mod(pt2_demo())

cell <- pt2_cell(mod$patterns[[1]], 0L, 0L)

pt2_note(cell)
}
