% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{pt2_render_options}
\alias{pt2_render_options}
\title{Retrieve options for rendering}
\usage{
pt2_render_options(...)
}
\arguments{
\item{...}{Specify custom options.}
}
\description{
Retrieve options for rendering Protracker modules. See also
\code{\link[=pt2_render]{pt2_render()}}.
}
\examples{
pt2_render_options(stereo_separation = 100)
}
\author{
Pepijn de Vries
}
