% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{pt2_sample}
\alias{pt2_sample}
\title{Obtain sample data and info from a ProTracker module}
\usage{
pt2_sample(mod, i, ...)
}
\arguments{
\item{mod}{An object of class \code{pt2mod} from which to obtain sample data and information}

\item{i}{The index of the requested sample (between 0 and 30).}

\item{...}{Ignored.}
}
\value{
Returns a sample object of class \code{pt2samp}.
}
\description{
Obtain sample data and info from a ProTracker module at a specific index.
ProTracker modules can hold up to 31 samples. The index should range from 0 to 30.
}
\examples{
mod <- pt2_read_mod(system.file("mod.intro", package = "ProTrackR2"))

smp <- pt2_sample(mod, 0L)
}
\author{
Pepijn de Vries
}
