% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ops.R
\name{$.pt2mod}
\alias{$.pt2mod}
\alias{$<-.pt2mod}
\alias{[[.pt2mod}
\alias{[.pt2patlist}
\alias{[[.pt2patlist}
\alias{[[<-.pt2patlist}
\alias{[.pt2samplist}
\alias{[[.pt2samplist}
\alias{[.pt2pat}
\alias{[<-.pt2pat}
\alias{[[.pt2celllist}
\alias{[.pt2celllist}
\alias{[[.pt2command}
\alias{[.pt2command}
\title{Select and assign operators for ProTrackR2 S3 class objects}
\usage{
\method{$}{pt2mod}(x, i, ...)

\method{$}{pt2mod}(x, i) <- value

\method{[[}{pt2mod}(x, i, ...)

\method{[}{pt2patlist}(x, i, ...)

\method{[[}{pt2patlist}(x, i, ...)

\method{[[}{pt2patlist}(x, i) <- value

\method{[}{pt2samplist}(x, i, ...)

\method{[[}{pt2samplist}(x, i, ...)

\method{[}{pt2pat}(x, i, j, ...)

\method{[}{pt2pat}(x, i, j, ...) <- value

\method{[[}{pt2celllist}(x, i, ...)

\method{[}{pt2celllist}(x, i, ...)

\method{[[}{pt2command}(x, i, ...)

\method{[}{pt2command}(x, i, ...)
}
\arguments{
\item{x}{Object to apply S3 method to. See 'usage' section for allowed object types.}

\item{i, j}{Indices for extracting or replacing ProTrackR2 object elements}

\item{...}{Passed on to other methods.}

\item{value}{Replacement value for the selected element(s).}
}
\value{
Returns the selected object in case of a selection (\code{[}, \code{[[}, or \code{$}) operator. Returns the
updated object \code{x} in case of an assignment (\verb{<-}) operator.
}
\description{
Functions to select and assign elements to ProTracker modules.
See \code{vignette('s3class')} for an overview of ProTrackR2 S3 class objects. See \code{vignette('sel_assign')}
for practical guidance on selecting and assigning elements of ProTrackR2 class objects.
}
