% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{play}
\alias{play}
\alias{play.pt2mod}
\title{Play a Protracker module}
\usage{
\method{play}{pt2mod}(x, duration = 120, ...)
}
\arguments{
\item{x}{Object to be played.}

\item{duration}{Duration in seconds to play.}

\item{...}{Arguments passed to \code{\link[=pt2_render]{pt2_render()}}.}
}
\value{
Returns an \verb{[audio::}$.audioInstance\verb{]} object which
allows you to control the playback (pause, resume, rewind).
}
\description{
Renders a Protracker module as \code{\link[audio:audioSample]{audio::audioSample()}} and plays it.
}
\examples{
\dontrun{
mod <- pt2_read_mod(system.file("mod.intro", package = "ProTrackR2"))

## ctrl will contain the audioInstance that will let you control the audio playback:
ctrl <- play(mod)
}
}
\author{
Pepijn de Vries
}
