% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{pt2_instrument}
\alias{pt2_instrument}
\title{Extract a sample index from a ProTracker pattern}
\usage{
pt2_instrument(x, ...)
}
\arguments{
\item{x}{An object of class \code{pt2_cell}, which can be extracted
from a pattern table with \code{\link[=pt2_cell]{pt2_cell()}}.}

\item{...}{Ignored.}
}
\value{
Returns the \code{integer} sample index in \code{x}. The index
has a base of \code{1}. An index of \code{0} means 'no sample'.
}
\description{
As explained in \code{vignette("s3class")}, the ProTracker pattern table
consists of cells containing information about the note and instrument
to be played. This function extracts the sample index (instrument)
from such a cell.
}
\examples{
mod <- pt2_read_mod(pt2_demo())

cell <- pt2_cell(mod$patterns[[1]], 0L, 0L)

pt2_instrument(cell)
}
