% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pt2_pattern}
\alias{pt2_pattern}
\title{Retrieve a pattern from a Protracker module}
\usage{
pt2_pattern(mod, i, ...)
}
\arguments{
\item{mod}{A \code{pt2mod} class objects from which to retrieve a pattern table}

\item{i}{The index (\code{integer}) of the pattern. Note that the index starts at 0.}

\item{...}{Ignored}
}
\value{
A \code{pt2pat} object representing the pattern.
}
\description{
Get a pattern table (sequence of notes and effects on each of the 4 channels) at
a specific index from a Protracker module.
}
\examples{
mod <- pt2_read_mod(system.file("mod.intro", package = "ProTrackR2"))

pat <- pt2_pattern(mod, 0L)
}
\author{
Pepijn de Vries
}
