% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{pt2_render}
\alias{pt2_render}
\alias{pt2_render.pt2mod}
\title{Render Protracker modules to a playable format}
\usage{
pt2_render(x, duration = 120, ...)

\method{pt2_render}{pt2mod}(x, duration = 120, ...)
}
\arguments{
\item{x}{The object to be rendered}

\item{duration}{Duration of the rendered output in seconds.}

\item{...}{Ignored}
}
\value{
Rendered audio inheriting the \code{\link[audio:audioSample]{audio::audioSample()}} class.
}
\description{
Renders a 16bit pulse-code modulation waveform from a Protracker module.
The rendered format can be played on a modern machine.
}
\examples{
mod <- pt2_read_mod(system.file("mod.intro", package = "ProTrackR2"))
aud <- pt2_render(mod)
}
\author{
Pepijn de Vries
}
